Moved location of files for sh1106 commands
